+++ src/App.jsx
@@ -25,7 +25,7 @@ import AdminDashboard from "../src/views/AdminDashboard/AdminDashboard.jsx";
 import RequestPasswordReset from "./views/Password/RequestPasswordReset";
 import PasswordReset from "./views/Password/PasswordReset";
 import UserAdmin from "../src/views/UserAdminDashboard/UserAdmin.jsx";
-import CreateCouponForm from "./views/AdminDashboard/FormsAdmDash.jsx/reateCouponForm.jsx";
+import CreateCouponForm from "./views/AdminDashboard/FormsAdmDash/CreateCouponForm.jsx";
 import NotFound404 from "./views/NotFound404/NotFound404.jsx";
 
 function App() {
@@ -80,7 +80,7 @@ function App() {
               element={<PasswordReset />}
             />
             <Route path="/crear-cupon" element={<CreateCouponForm />} />
-            <Route path='*' element={<NotFound404 />}/>
+            <Route path="*" element={<NotFound404 />} />
           </Routes>
         </div>
       </>
